requirements.txt
django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.5
python-decouple==3.8
django-debug-toolbar==4.2.1
